reorganize file structure after rebase
