+++ src/main/java/com/odos/odos_server/domain/challenge/entity/Challenge.java
@@ -3,7 +3,6 @@ package com.odos.odos_server.domain.challenge.entity;
 import com.odos.odos_server.domain.common.Enum.ChallengeCategory;
 import com.odos.odos_server.domain.common.Enum.ChallengeType;
 import com.odos.odos_server.domain.diary.entity.Diary;
-import com.odos.odos_server.domain.member.entity.Member;
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
 import java.util.List;
