+++ src/main/java/com/odos/odos_server/domain/challenge/entity/ChallengeLike.java
@@ -1,6 +1,5 @@
 package com.odos.odos_server.domain.challenge.entity;
 
-import com.odos.odos_server.domain.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
