+++ src/main/java/com/odos/odos_server/domain/common/Enum/MemberRole.java
@@ -1,5 +1,8 @@
 package com.odos.odos_server.domain.common.Enum;
 
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
 @Getter
 @RequiredArgsConstructor
 public enum MemberRole {
