+++ src/main/java/com/odos/odos_server/domain/common/Enum/SignupRoute.java
@@ -1,8 +1,13 @@
 package com.odos.odos_server.domain.common.Enum;
 
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
 public enum SignupRoute {
   KAKAO,
+  GOOGLE,
   NAVER,
-  APPLE,
-  GOOGLE
+  APPLE
 }
