+++ src/main/java/com/odos/odos_server/domain/diary/entity/Diary.java
@@ -1,8 +1,7 @@
 package com.odos.odos_server.domain.diary.entity;
 
-import com.odos.odos_server.domain.common.Enum.Feeling;
 import com.odos.odos_server.domain.challenge.entity.Challenge;
-import com.odos.odos_server.domain.member.entity.Member;
+import com.odos.odos_server.domain.common.Enum.Feeling;
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
 import java.util.List;
