+++ src/main/java/com/odos/odos_server/domain/diary/entity/DiaryReport.java
@@ -1,7 +1,6 @@
 package com.odos.odos_server.domain.diary.entity;
 
 import com.odos.odos_server.domain.common.Enum.ReportType;
-import com.odos.odos_server.domain.member.entity.Member;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
