+++ src/main/java/com/odos/odos_server/domain/member/dto/SignupInfoRequest.java
@@ -1,7 +1,7 @@
-package com.odos.odos_server.member.dto;
+package com.odos.odos_server.domain.member.dto;
 
-import com.odos.odos_server.member.enums.Gender;
-import com.odos.odos_server.member.enums.Job;
+import com.odos.odos_server.domain.common.Enum.Gender;
+import com.odos.odos_server.domain.common.Enum.Job;
 import java.time.LocalDate;
 import lombok.Getter;
 import org.antlr.v4.runtime.misc.NotNull;
