--- src/main/java/com/odos/odos_server/member/enums/Gender.java
@@ -1,7 +1,0 @@
-package com.odos.odos_server.member.enums;
-
-public enum Gender {
-  MALE,
-  FEMALE,
-  ETC
-}
