--- src/main/java/com/odos/odos_server/member/enums/Job.java
@@ -1,7 +1,0 @@
-package com.odos.odos_server.member.enums;
-
-public enum Job {
-  STUDENT,
-  EMPLOYEE,
-  ETC
-}
