--- src/main/java/com/odos/odos_server/member/enums/MemberRole.java
@@ -1,14 +1,0 @@
-package com.odos.odos_server.member.enums;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum MemberRole {
-  USER("ROLE_USER"),
-  ADMIN("ROLE_ADMIN"),
-  GUEST("ROLE_GUEST");
-
-  private final String key;
-}
