--- src/main/java/com/odos/odos_server/member/enums/SignupRoute.java
@@ -1,13 +1,0 @@
-package com.odos.odos_server.member.enums;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum SignupRoute {
-  KAKAO,
-  GOOGLE,
-  NAVER,
-  APPLE
-}
